Moved not so common files out of common-utils to their respective modules
