+++ plugin.maven/src/main/java/com/phsyberdome/plugin/maven/MavenVersionHelper.java
@@ -1,7 +1,8 @@
 
 
-package com.phsyberdome.common.utils;
+package com.phsyberdome.plugin.maven;
 
+import com.phsyberdome.common.utils.CLIHelper;
 import com.vdurmont.semver4j.Semver;
 import com.vdurmont.semver4j.SemverException;
 import java.util.ArrayList;
